Initial extraction of TickerSearchModal
