+++ src/hooks/useSearch.ts
@@ -25,6 +25,7 @@ const DEFAULT_PROPS: Required<UseSearchProps> = {
 };
 
 // TODO: Capture error state
+// TODO: Rename to `useTickerSearch`
 export default function useSearch(
   props: Partial<UseSearchProps> = DEFAULT_PROPS,
 ) {
